+++ package.json
@@ -206,7 +206,7 @@
     "@lobehub/icons": "^4.0.2",
     "@lobehub/market-sdk": "^0.25.1",
     "@lobehub/tts": "^4.0.2",
-    "@lobehub/ui": "^4.6.3",
+    "@lobehub/ui": "^4.6.6",
     "@modelcontextprotocol/sdk": "^1.25.1",
     "@neondatabase/serverless": "^1.0.2",
     "@next/third-parties": "^16.1.1",
