+++ src/features/Conversation/Messages/Task/Actions/index.tsx
@@ -1,9 +1,12 @@
 import { type UIChatMessage } from '@lobechat/types';
-import { ActionIconGroup } from '@lobehub/ui';
+import { ActionIconGroup, createRawModal } from '@lobehub/ui';
 import type { ActionIconGroupEvent, ActionIconGroupItemType } from '@lobehub/ui';
-import { memo, useCallback, useMemo, useState } from 'react';
+import { memo, useCallback, useMemo } from 'react';
 
-import ShareMessageModal from '../../../components/ShareMessageModal';
+import { useEventCallback } from '@/hooks/useEventCallback';
+
+import ShareMessageModal, { type ShareModalProps } from '../../../components/ShareMessageModal';
+import { Provider, createStore, useConversationStoreApi } from '../../../store';
 import type {
   MessageActionItem,
   MessageActionItemOrDivider,
@@ -56,13 +59,35 @@ interface AssistantActionsBarProps {
 export const AssistantActionsBar = memo<AssistantActionsBarProps>(
   ({ actionsConfig, id, data, index }) => {
     const { error, tools } = data;
-    const [showShareModal, setShareModal] = useState(false);
+    const store = useConversationStoreApi();
+    const handleOpenShareModal = useEventCallback(() => {
+      createRawModal(
+        (props: ShareModalProps) => (
+          <Provider
+            createStore={() => {
+              const state = store.getState();
+              return createStore({
+                context: state.context,
+                hooks: state.hooks,
+                skipFetch: state.skipFetch,
+              });
+            }}
+          >
+            <ShareMessageModal {...props} />
+          </Provider>
+        ),
+        {
+          message: data,
+        },
+        { onCloseKey: 'onCancel', openKey: 'open' },
+      );
+    });
 
     const defaultActions = useAssistantActions({
       data,
       id,
       index,
-      onOpenShareModal: () => setShareModal(true),
+      onOpenShareModal: handleOpenShareModal,
     });
 
     const hasTools = !!tools;
@@ -159,16 +184,7 @@ export const AssistantActionsBar = memo<AssistantActionsBarProps>(
 
     if (error) return <ErrorActionsBar actions={defaultActions} onActionClick={handleAction} />;
 
-    return (
-      <>
-        <ActionIconGroup items={items} menu={{ items: menu }} onActionClick={handleAction} />
-        <ShareMessageModal
-          message={data}
-          onCancel={() => setShareModal(false)}
-          open={showShareModal}
-        />
-      </>
-    );
+    return <ActionIconGroup items={items} menu={{ items: menu }} onActionClick={handleAction} />;
   },
 );
