+++ src/features/Conversation/Messages/index.tsx
@@ -4,29 +4,13 @@ import { isDesktop } from '@lobechat/const';
 import { Flexbox } from '@lobehub/ui';
 import { createStaticStyles, cx } from 'antd-style';
 import isEqual from 'fast-deep-equal';
-import {
-  type MouseEvent,
-  type ReactNode,
-  type RefObject,
-  Suspense,
-  memo,
-  useCallback,
-  useMemo,
-  useRef,
-} from 'react';
-import type { VListHandle } from 'virtua';
+import { type MouseEvent, type ReactNode, Suspense, memo, useCallback } from 'react';
 
 import BubblesLoading from '@/components/BubblesLoading';
 
-import ContextMenu from '../components/ContextMenu';
 import History from '../components/History';
 import { useChatItemContextMenu } from '../hooks/useChatItemContextMenu';
-import {
-  dataSelectors,
-  messageStateSelectors,
-  useConversationStore,
-  virtuaListSelectors,
-} from '../store';
+import { dataSelectors, messageStateSelectors, useConversationStore } from '../store';
 import AgentCouncilMessage from './AgentCouncil';
 import AssistantMessage from './Assistant';
 import AssistantGroupMessage from './AssistantGroup';
@@ -73,13 +57,10 @@ const MessageItem = memo<MessageItemProps>(
     index,
     isLatestItem,
   }) => {
-    const containerRef = useRef<HTMLDivElement | null>(null);
-
     const topic = useConversationStore((s) => s.context.topicId);
 
-    // Get message and actionsBar from ConversationStore
+    // Get message from ConversationStore
     const message = useConversationStore(dataSelectors.getDisplayMessageById(id), isEqual);
-    const actionsBar = useConversationStore((s) => s.actionsBar);
     const role = message?.role;
 
     const [editing, isMessageCreating] = useConversationStore((s) => [
@@ -87,23 +68,12 @@ const MessageItem = memo<MessageItemProps>(
       messageStateSelectors.isMessageCreating(id)(s),
     ]);
 
-    const {
-      containerRef: contextMenuContainerRef,
-      contextMenuState,
-      handleContextMenu,
-      hideContextMenu,
-    } = useChatItemContextMenu({
+    const { handleContextMenu } = useChatItemContextMenu({
       editing,
-      onActionClick: () => {},
+      id,
+      inPortalThread,
+      topic,
     });
-
-    const setContainerRef = useCallback(
-      (node: HTMLDivElement | null) => {
-        containerRef.current = node;
-        contextMenuContainerRef.current = node;
-      },
-      [contextMenuContainerRef],
-    );
 
     const onContextMenu = useCallback(
       async (event: MouseEvent<HTMLDivElement>) => {
@@ -128,15 +98,6 @@ const MessageItem = memo<MessageItemProps>(
         handleContextMenu(event);
       },
       [handleContextMenu, id, role, message],
-    );
-    // Get virtuaScrollMethods from ConversationStore
-    const virtuaScrollMethods = useConversationStore(virtuaListSelectors.virtuaScrollMethods);
-
-    // Create a ref-like object for ContextMenu compatibility
-    // VirtuaScrollMethods is a subset of VListHandle with the methods ContextMenu needs
-    const virtuaRef = useMemo<RefObject<VListHandle | null>>(
-      () => ({ current: virtuaScrollMethods as VListHandle | null }),
-      [virtuaScrollMethods],
     );
 
     const renderContent = useCallback(() => {
@@ -202,7 +163,7 @@ const MessageItem = memo<MessageItemProps>(
       }
 
       return null;
-    }, [role, disableEditing, id, index, isLatestItem, actionsBar]);
+    }, [role, disableEditing, id, index, isLatestItem]);
 
     if (!role) return;
 
@@ -213,22 +174,10 @@ const MessageItem = memo<MessageItemProps>(
           className={cx(styles.message, className, isMessageCreating && styles.loading)}
           data-index={index}
           onContextMenu={onContextMenu}
-          ref={setContainerRef}
         >
           <Suspense fallback={<BubblesLoading />}>{renderContent()}</Suspense>
           {endRender}
         </Flexbox>
-        <ContextMenu
-          id={id}
-          inPortalThread={inPortalThread}
-          index={index}
-          onClose={hideContextMenu}
-          position={contextMenuState.position}
-          selectedText={contextMenuState.selectedText}
-          topic={topic}
-          virtuaRef={virtuaRef}
-          visible={contextMenuState.visible}
-        />
       </>
     );
   },
