+++ src/features/ResourceManager/components/Explorer/ToolBar/SortDropdown.tsx
@@ -1,5 +1,4 @@
-import { Dropdown } from 'antd';
-import type { MenuProps } from 'antd';
+import { type DropdownItem, DropdownMenu } from '@lobehub/ui';
 import { ArrowDownAZ } from 'lucide-react';
 import { memo, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -22,22 +21,23 @@ const SortDropdown = memo(() => {
     [t],
   );
 
-  const menuItems: MenuProps['items'] = sortOptions.map((option) => ({
+  const menuItems: DropdownItem[] = sortOptions.map((option) => ({
     key: option.key,
     label: option.label,
     onClick: () => setSorter(option.key as 'name' | 'createdAt' | 'size'),
+    style:
+      option.key === (sorter || 'createdAt')
+        ? { backgroundColor: 'var(--ant-control-item-bg-active)' }
+        : {},
   }));
 
   const currentSortLabel =
     sortOptions.find((option) => option.key === sorter)?.label || t('FileManager.sort.dateAdded');
 
   return (
-    <Dropdown
-      menu={{ items: menuItems, selectedKeys: [sorter || 'createdAt'] }}
-      trigger={['click']}
-    >
+    <DropdownMenu items={menuItems}>
       <ActionIconWithChevron icon={ArrowDownAZ} title={currentSortLabel} />
-    </Dropdown>
+    </DropdownMenu>
   );
 });
