+++ src/features/ResourceManager/components/Explorer/ToolBar/ViewSwitcher.tsx
@@ -1,6 +1,4 @@
-import { Icon } from '@lobehub/ui';
-import { Dropdown } from 'antd';
-import type { MenuProps } from 'antd';
+import { type DropdownItem, DropdownMenu, Icon } from '@lobehub/ui';
 import { Grid3x3Icon, ListIcon } from 'lucide-react';
 import { memo, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -20,31 +18,30 @@ const ViewSwitcher = memo(() => {
   const currentViewLabel =
     viewMode === 'list' ? t('FileManager.view.list') : t('FileManager.view.masonry');
 
-  const menuItems = useMemo<MenuProps['items']>(() => {
+  const menuItems = useMemo<DropdownItem[]>(() => {
     return [
       {
         icon: <Icon icon={ListIcon} />,
         key: 'list',
         label: t('FileManager.view.list'),
         onClick: () => setViewMode('list'),
+        style: viewMode === 'list' ? { backgroundColor: 'var(--ant-control-item-bg-active)' } : {},
       },
       {
         icon: <Icon icon={Grid3x3Icon} />,
         key: 'masonry',
         label: t('FileManager.view.masonry'),
         onClick: () => setViewMode('masonry'),
+        style:
+          viewMode === 'masonry' ? { backgroundColor: 'var(--ant-control-item-bg-active)' } : {},
       },
     ];
-  }, [setViewMode, t]);
+  }, [setViewMode, t, viewMode]);
 
   return (
-    <Dropdown
-      menu={{ items: menuItems, selectedKeys: [viewMode] }}
-      placement="bottomRight"
-      trigger={['click']}
-    >
+    <DropdownMenu items={menuItems} placement="bottomRight">
       <ActionIconWithChevron icon={currentViewIcon} title={currentViewLabel} />
-    </Dropdown>
+    </DropdownMenu>
   );
 });
